Incorporate post-migration markup updates into MarkdownEngine
