+++ src/main/java/io/pivotal/util/MarkupEngine.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2016 the original author or authors.
+ * Copyright 2002-2023 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -34,8 +34,14 @@ public interface MarkupEngine {
 		// no-op
 	}
 
-	String link(String text, String href);
+	/**
+	 * Format the given description and URL according to the markup.
+	 */
+	String link(String description, String url);
 
+	/**
+	 * Convert the text from JIRA to the target markup.
+	 */
 	String convert(String text);
 
 }
